Page-Agile/Workspace-reservation: move into folder && refactor styling constant
