+++ src/navigators/AgileOfficeTab.tsx
@@ -5,8 +5,8 @@ import { Image, useColorModeValue } from "native-base";
 import AgileOfficeMain from "screens/agileOffice";
 import { AgileOfficeCLoc } from "screens/agileOffice/AgileOfficeCLoc";
 import { AgileOfficeScanQR } from "screens/agileOffice/AgileOfficeScanQR";
-import { AgileOfficeWorkspaceReservation } from "screens/agileOffice/AgileOfficeWorkspaceReservation";
-import { AgileOfficeWorkspaceReservationSearchResult } from "screens/agileOffice/AgileOfficeWorkspaceReservationSearchResult";
+import { AgileOfficeWorkspaceReservation } from "screens/agileOffice/screens/AgileOfficeWorkspaceReservation";
+import { AgileOfficeWorkspaceReservationSearchResult } from "screens/agileOffice/screens/AgileOfficeWorkspaceReservationSearchResult";
 import { AgileOfficeScanResult } from "screens/agileOffice/AgileOfficeScanResult";
 import { AgileOfficeAreaBook } from "screens/agileOffice/AgileOfficeAreaBook";
 import { header_light, unifi_c1, unifi_c4, unifi_c7 } from "components/styles";
